Revert "location refactoring (#226)"
This reverts commit 5c5b8adb02c2129f1c6a571bcb3f96526b4dfb86.
