+++ TextToSpeech.Web/src/app/components/tts-form/tts-form.component.ts
@@ -9,7 +9,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { Subscription } from 'rxjs';
-import { EnLanguageCode } from '../../constants/content/language';
+import { EnLanguageCode } from '../../constants/language';
 import { TtsApis, Narakeet, DemoText } from '../../constants/tts-constants';
 import { SpeechClient } from '../../http-clients/speech-client';
 import { DropdownConfig } from '../../models/dropdown-config';
@@ -18,8 +18,8 @@ import { AudioService } from '../../services/audio.service';
 import { ConfigService } from '../../services/config.service';
 import { SignalRService } from '../../services/signalr.service';
 import { TranslationService } from '../../services/translation.service';
-import { DropdownComponent } from '../dropdown/dropdown.component';
-import { SnackbarService } from '../../ui-services/snackbar-service';
+import { DropdownComponent } from '../../shared-ui/components/dropdown/dropdown.component';
+import { SnackbarService } from '../../shared-ui/snackbar-service';
 import { DropdownService } from '../../services/dropdown.service';
 import { FileInputService } from '../../services/file-input.service';
 import { AcceptableFileTypes } from "../../constants/tts-constants";
