+++ TextToSpeech.Web/src/app/shared-ui/components/dropdown/dropdown.component.scss
